refactor(cli): new package name
