+++ packages/disploy/cli/src/types.ts
@@ -1,4 +1,4 @@
-import type { App, Command } from '@disploy/framework';
+import type { App, Command } from 'disploy';
 
 export interface DisployStandaloneBundle {
 	app: App;
